+++ clx_invoice_policy/models/sale_order.py
@@ -17,44 +17,44 @@ class SaleOrder(models.Model):
     management_company_id = fields.Many2one(related="partner_id.management_company_type_id", store=True)
     ownership_company_id = fields.Many2one(related="partner_id.ownership_company_type_id", store=True)
 
-    def _action_confirm(self):
-        """
-        Create draft invoice when confirm the sale order.
-        if sale order line start date is in current month.
-        than create invoice current month.
-        :return:
-        """
-        res = super(SaleOrder, self)._action_confirm()
-        if self.is_ratio:
-            return res
-        if self.subscription_management in ('upsell', 'downsell'):
-            return res
-        so_lines = self.env['sale.subscription.line'].search([
-            ('so_line_id.order_id', '=', self.id),
-        ])
-        current_month_start_day = fields.Date.today()
-        end_date = current_month_start_day + relativedelta(months=self.clx_invoice_policy_id.num_of_month + 1)
-        end_date = end_date - relativedelta(days=1)
-        lines = so_lines.filtered(lambda x: x.start_date and x.start_date < end_date)
-        count = self.clx_invoice_policy_id.num_of_month + 1
-        if lines:
-            so_lines = lines.filtered(lambda x: x.invoice_start_date.month == current_month_start_day.month)
-            for i in range(0, count):
-                if self.partner_id.child_invoice_selection:
-                    if self.partner_id.child_invoice_selection == 'sol':
-                        self.partner_id.with_context(cofirm_sale=True, sol=True).generate_advance_invoice(so_lines)
-                    else:
-                        self.partner_id.with_context(cofirm_sale=True).generate_advance_invoice(so_lines)
-                if not self.partner_id.child_invoice_selection and self.partner_id.invoice_selection:
-                    if self.partner_id.invoice_selection == 'sol':
-                        self.partner_id.with_context(cofirm_sale=True, sol=True).generate_advance_invoice(so_lines)
-                    else:
-                        self.partner_id.with_context(cofirm_sale=True).generate_advance_invoice(so_lines)
-                so_lines = lines.filtered(lambda x: (not x.end_date and x.invoice_start_date and x.invoice_start_date < end_date)
-                                                    or
-                                                    (x.end_date and x.invoice_start_date and x.invoice_start_date < end_date)
-                                          )
-        return res
+    # def _action_confirm(self):
+    #     """
+    #     Create draft invoice when confirm the sale order.
+    #     if sale order line start date is in current month.
+    #     than create invoice current month.
+    #     :return:
+    #     """
+    #     res = super(SaleOrder, self)._action_confirm()
+    #     if self.is_ratio:
+    #         return res
+    #     if self.subscription_management in ('upsell', 'downsell'):
+    #         return res
+    #     so_lines = self.env['sale.subscription.line'].search([
+    #         ('so_line_id.order_id', '=', self.id),
+    #     ])
+    #     current_month_start_day = fields.Date.today()
+    #     end_date = current_month_start_day + relativedelta(months=self.clx_invoice_policy_id.num_of_month + 1)
+    #     end_date = end_date - relativedelta(days=1)
+    #     lines = so_lines.filtered(lambda x: x.start_date and x.start_date < end_date)
+    #     count = self.clx_invoice_policy_id.num_of_month + 1
+    #     if lines:
+    #         so_lines = lines.filtered(lambda x: x.invoice_start_date.month == current_month_start_day.month)
+    #         for i in range(0, count):
+    #             if self.partner_id.child_invoice_selection:
+    #                 if self.partner_id.child_invoice_selection == 'sol':
+    #                     self.partner_id.with_context(cofirm_sale=True, sol=True).generate_advance_invoice(so_lines)
+    #                 else:
+    #                     self.partner_id.with_context(cofirm_sale=True).generate_advance_invoice(so_lines)
+    #             if not self.partner_id.child_invoice_selection and self.partner_id.invoice_selection:
+    #                 if self.partner_id.invoice_selection == 'sol':
+    #                     self.partner_id.with_context(cofirm_sale=True, sol=True).generate_advance_invoice(so_lines)
+    #                 else:
+    #                     self.partner_id.with_context(cofirm_sale=True).generate_advance_invoice(so_lines)
+    #             so_lines = lines.filtered(lambda x: (not x.end_date and x.invoice_start_date and x.invoice_start_date < end_date)
+    #                                                 or
+    #                                                 (x.end_date and x.invoice_start_date and x.invoice_start_date < end_date)
+    #                                       )
+    #     return res
 
     @api.onchange('partner_id')
     def onchange_partner_id(self):