+++ RELEASE_NOTES.md
@@ -1173,6 +1173,20 @@
 #### Version 13.1.1.0.24
 - [FIX] clx_subscription_creation : set description on upsell.
 
-## 01/18/2021
+## 01/22/2021
 #### Version 13.3.0.0.28
-- [FIX] clx_budget_analysis_report : fix the issue related to wholesale price now taken total of all subscription lines.
+- [FIX] clx_budget_analysis_report : fix the issue wholeprice and downsell.
+
+## 01/17/2021
+#### Version 13.1.0.0.19
+- [FIX] clx_retail_pricelist : hide the management fees and wholesale price on sale order.
+
+## 01/22/2021
+#### Version 13.1.0.0.60
+- [ADD] clx_invoice_policy : added code for the not create invoice for imported data for the past.
+
+## 01/22/2021
+#### Version 13.1.1.0.24
+- [FIX] clx_subscription_creation : Fix the issue related to description on subscription kanban view.
+
+